+++ yp-online/src/main/java/kr/co/yourplanet/online/properties/EncryptProperties.java
@@ -1,4 +1,4 @@
-package kr.co.yourplanet.online.common.encrypt;
+package kr.co.yourplanet.online.properties;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
